RTU UI update
